New event functions
A function that gives a  complete new save if something is wrong with the text file, and a function that adds a single event to the event to the text file. Both functions take in a single event array. Also a function that reads the text file and returns its contents in a string.